Start the expanding animation immediately
The expanding animation would start only after the first `speed`
milliseconds have expired, after which it might be already too late for
it to start. There is no reason to wait for the animation to start once
the `isLoading` property is detected to be `true`, which is what this
commit does.